rename: FeaturePage.tsx → Feature.tsx
